+++ src/components/index.ts
@@ -3,3 +3,4 @@ export { default as SinglePageProduct } from "./single-page-product";
 export { default as Navbar } from "./navbar";
 export { default as QuantityInput } from "./quantity-input";
 export { default as CartItem } from "./cart-item";
+export { default as EmptyCart } from "./empty-cart";
